Hotfix 1.2.2

+++ src/UdpHubListener.h
@@ -104,7 +104,7 @@ private:
    */
     static void bindUdpSocket(QUdpSocket& udpsocket, int port);
 
-    int readClientUdpPort(QTcpSocket* clientConnection);
+    uint16_t readClientUdpPort(QTcpSocket* clientConnection);
     int sendUdpPort(QTcpSocket* clientConnection, int udp_port);
